requirements.txt
fastapi==0.92.0
fitz==0.0.1.dev2
Pillow==9.4.0
PyMuPDF==1.21.1
